+++ backend/urls.py
@@ -19,6 +19,6 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('api/', include('backend.reviews.urls')),
+    path("api/", include("backend.api.urls")),
     path("auth/", include("backend.auth.urls")),
 ]